cleanup

+++ src/FM.LiveSwitch.Connect/RtpPacket.cs
@@ -3,13 +3,13 @@
     class RtpPacket
     {
         public DataBuffer Payload { get; set; }
-        public long SequenceNumber { get; set; }
+        public int SequenceNumber { get; set; }
         public long Timestamp { get; set; }
         public bool Marker { get; set; }
         public int PayloadType { get; set; }
         public long SynchronizationSource { get; set; }
 
-        public RtpPacket(DataBuffer payload, long sequenceNumber, long timestamp, bool marker)
+        public RtpPacket(DataBuffer payload, int sequenceNumber, long timestamp, bool marker)
         {
             Payload = payload;
             SequenceNumber = sequenceNumber;
